Backend/requirements.txt
fastapi
uvicorn[standard]

pydantic
python-dotenv
python-multipart
email-validator
passlib[bcrypt]
bcrypt
python-jose

requests
httpx

pymongo

numpy
pandas
scikit-learn
joblib

tensorflow-cpu
keras
h5py

google-generativeai

protobuf>=4.25.8,<6

tqdm
rich
pillow
